device-pi/requirements.txt
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
requests==2.31.0
eth-account==0.10.0
python-dotenv==1.0.0
qrcode[pil]==7.4.2
supabase==2.3.0